malayalam-telegram-bot/requirement.txt
python-telegram-bot==20.6
gspread==6.0.2
oauth2client==4.1.3
